added chargable interface

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/Weighable.java
@@ -3,6 +3,6 @@ package kenakata.catalog;
 public interface Weighable {
 
 
-long weightGrams();
+double weightGrams();
 
 }
